Removed abstract container, introduced composite container
