+++ src/FactoryInterface.php
@@ -7,7 +7,6 @@
 
 namespace yii\di;
 
-use Psr\Container\ContainerInterface;
 use yii\di\exceptions\InvalidConfigException;
 
 /**
