+++ src/dependencies/ClassDependency.php
@@ -8,7 +8,6 @@
 namespace yii\di\dependencies;
 
 use Psr\Container\ContainerInterface;
-use yii\di\AbstractContainer;
 use yii\di\contracts\DependencyInterface;
 use yii\di\exceptions\InvalidConfigException;
 
@@ -31,7 +30,7 @@ class ClassDependency implements DependencyInterface
         $this->optional = $optional;
     }
 
-    public function resolve(AbstractContainer $container)
+    public function resolve(ContainerInterface $container)
     {
         try {
             $result = $container->get($this->class);
