+++ src/dependencies/InvalidDependency.php
@@ -8,7 +8,6 @@
 namespace yii\di\dependencies;
 
 use Psr\Container\ContainerInterface;
-use yii\di\AbstractContainer;
 use yii\di\contracts\DependencyInterface;
 use yii\di\exceptions\NotFoundException;
 
@@ -23,7 +22,7 @@ use yii\di\exceptions\NotFoundException;
  */
 class InvalidDependency implements DependencyInterface
 {
-    public function resolve(AbstractContainer $container)
+    public function resolve(ContainerInterface $container)
     {
         throw new NotFoundException('Invalid reference');
     }
